+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.2.0
 
+- feat: add more parameters to `ProgressBar`
 - **BREAKING**: Change `moveForward` to `moveRight`, `moveBackward` to `moveLeft`
 - **BREAKING**: Change `clearCurrentToEndScreen` to `clearCurrentToScreenEnd` (the same for other similar methods)